Añadir controladores y configurarlo todo correctamente
